requirements.txt
asgiref==3.7.2
attrs==24.2.0
az-iranian-bank-gateways==2.0.8
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
# django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-session-timeout==0.1.0
django-widget-tweaks==1.5.0
dnspython==2.6.1
email_validator==2.2.0
idna==3.6
isodate==0.6.1
lxml==5.3.0
phonenumbers==8.13.43
Pillow==10.1.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
zeep==4.2.1
